Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    int num;
+    float num;
     printf("num is",num);
-    scanf("%f",&num);
+    scanf("%.1f",&num);
     return 0;
 }
